# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create/category": {
            "post": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request that added a new category",
                "description": "",
                "operationId": "c93ace596ff38f36f0b8a2f731f7495e",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Please write a category Photo File Name",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/update/category": {
            "post": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request which updating something regarding category",
                "description": "",
                "operationId": "83a1b93582b6e800d1b44d33959d752d",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Please write a category ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/category": {
            "post": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request which deletes category",
                "description": "",
                "operationId": "571a380100d6b27792d6010cd55fdfa5",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Please write a category ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/filter/catagory/byTitle": {
            "get": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request which returns categories with products, photos and options",
                "description": "",
                "operationId": "4347213db01f30a7265c1cbc341af173",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/catagories": {
            "get": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request which returns categories",
                "description": "",
                "operationId": "f2a2808c52a1e6bf77513b76346ca398",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/catagories/with/sub/catagories": {
            "get": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request which returns catagories with sub categories",
                "description": "",
                "operationId": "0878732fa440f6ed18125afd7d024e0d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
        "/api/create/sub/category": {
            "post": {
                "tags": [
                    "Sub Category Section"
                ],
                "summary": "Request that added a new sub category",
                "description": "",
                "operationId": "1d9b90c17470af22232a478403c3e6ca",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Please write a category Photo File Name",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/update/sub/category": {
            "post": {
                "tags": [
                    "Sub Category Section"
                ],
                "summary": "Request which updating something regarding Sub category",
                "description": "",
                "operationId": "68dd65a0872dfad9b99f290a4f46156d",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "sub_category_id",
                        "in": "query",
                        "description": "Please write a sub category ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/sub/category": {
            "post": {
                "tags": [
                    "Sub Category Section"
                ],
                "summary": "Request which deletes sub category",
                "description": "",
                "operationId": "28c71de1cf9813529e162725ec38448e",
                "parameters": [
                    {
                        "name": "sub_category_id",
                        "in": "query",
                        "description": "Please write a sub category ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}